+++ CHANGELOG.md
@@ -5,7 +5,8 @@
    * add [pdf test asset and pdf test](https://github.com/iambumblehead/thu.sh/pull/58)
    * add [video test asset and video test](https://github.com/iambumblehead/thu.sh/pull/58)
    * add [iTerm2 detection](https://github.com/iambumblehead/thu.sh/pull/60)
-   * add [font test asset and test](https://github.com/iambumblehead/thu.sh/pull/61)
+   * add [font test asset and test](https://github.com/iambumblehead/thu.sh/pull/62)
+   * add [audio test asset and test](https://github.com/iambumblehead/thu.sh/pull/63)
  * 0.1.0 _Mar.27.2024_
    * remove [badgesize.io badge,](https://github.com/iambumblehead/thu.sh/pull/52) not working
    * use [integer timeout value](https://github.com/iambumblehead/thu.sh/pull/53) for darwin/mac read